Remove ZENKIT_VERSION definition from .env
Move the definition of ZENKIT_VERSION from its shared location in the
.env file into the Makefile and Dockerfile.

This change was initially motivated by the need to override images for
development when running the full stack in zing-deploy.  In that usage,
a local override file is created in zing-deploy which sets the image's
build context to the project's directory.  However, this approach does
not include the project's .env, so the ZENKIT_VERSION was not defined.
Moving the definition from the .env to the Dockerfile resolves this
problem.

There was some concern about duplicating the definition in both the
Dockerfile and the Makefile; however after consideration it became
apparent that the two usages are distinct: the one in the Makefile is
used as an image for running various go-related tools; the one in the
Dockerfile is a base image for building the image.